CIS Benchmarks Plugin
Currently publishing this and targeting a demo build I created myself
from the kube-bench. This is just so that we can publish this for
visibility while kube-bench finalizes pushing their v0.2.0 image.

That release/image is necessary so that it supports the junit output
which makes it nicer for Sonobuoy post-processing.

The main podSpec options for this were taken from the job.yaml file
in the aquasecurity/kube-bench repo. Various differences in cluster
settings/deployment method may require these to be tweaked slightly.

As people try out the plugin we will probably find ways to improve
the plugin so that it can support a wider array of clusters successfully.

Signed-off-by: John Schnake <jschnake@vmware.com>