Add kube-hunter plugin
Adds a plugin based on aquasecurity/kube-hunter.

References a image we built from kube-hunter source but users
can also use the aquasec/kube-hunter image or manually build their
own.

The reason for providing our own as a defult is because the aquasec
image contains closed source additions which ential accepting a EULA.
That may be fine for some users, but didn't want to default to that.

Signed-off-by: John Schnake <jschnake@vmware.com>