Add sleep at the end of kube-bench to avoid restart
Run as a daemonset, the restart policy is Always by mandate. As
a result, the kubelet sees the plugin completion as a call to restart
it. This leads to confusing messages/state in the logs because the
container ends up in a crashbackoffloop.

This PR just adds a sleep similar to how the systemd-logs plugin
does.

Fixes #10

Signed-off-by: John Schnake <jschnake@vmware.com>