Add CIS plugin support for Enterprise PKS
This adds Enterprise PKS specific support for the CIS benchmark. This
is required as Enterprise PKS uses different locations for various
configuration files which are not included within the standard
kube-bench configuration. Due to the fact that workloads cannot be run
on master nodes of Enterprise PKS clusters, only the worker node version
of the plugin has been provided.

It also renames the environment variable `PROVIDER` to `DISTRIBUTION`
and documents how to use this variable as it was previously unused.

As the configuration for kube-bench exists as a file on disk, the
configuration file for this version is added to the existing custom
image provided by the Sonobuoy team for running kube-bench.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>