Upgrade to latest version of kube-bench
This upgrades the version of kube-bench that we are using to latest
version, 0.2.3. This version includes support for CIS 1.15.

Several changes to the structure of the checks were introduced in thi
version, including the checks being separated into multiple "targets".
As a result, we would like to provide a way for the user to easily
customize which version and targets they run. This has been done by
introducing a small wrapper script which will check the value of
various plugin environment variables and call kube-bench accordingly.

To include this wrapper script, we need to provide our own version of
the kube-bench image with it included. This uses the Aqua Security image
as the base image and just adds the new script.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>