Add support for GKE (#19)
GKE v1.15+ clusters use a version of the CIS benchmark that is tailored
to GKE clusters. To use this, the flag `--benchmark` needs to be passed
to kube-bench rather than `--version`. This updates the script that runs
kube-bench to check for the GKE distribution. If the Kubernetes version
is set and is less than 1.15, it defaults to the original `--version`
flag, otherwise it assumes `--benchmark gke-1.0` should be used.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>