Only add extra targets if supported by K8s version (#22)
There are a number of targets which are only compatible if running
Kubernetes 1.15 and later (which means running CIS 1.5). If the version
is known by checking the value of `KUBERNETES_VERSION`, then only add
those targets which are supported by that version.

This also applies to the `managedservices` target for the GKE benchmark.
This target is only applicable if running the GKE benchmark which in
turn can only be run on Kubernetes 1.15 and later. Only add this target
if the cluster is GKE and K8s 1.15 or later.

If the K8s version is not known (`KUBERNETES_VERSION` is not set), then
all targets will be added if they are requested by the user.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>