Flatten results in sonobuoy report
By having nested results in the sonobuoy_results.yaml report, it meant
that a lot of the relevant details for each action check were missing
when using the `detailed` mode of `sonobuoy results`. Instead flatten
the results so that there is only one level of result items which have
all the details for what the subject can do.

In the `details` map, include the subject details, the action details
and the role bindings and cluster role bindings that allow that subject
to perform that action.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>