Don't run if no output file provided
Previously, if the user did not provide an output file, who-can would
still run but produce no output with no feedback provided to the user.
Now there is a pre-run hook that checks that at least one output file
has been specified.

This change also includes a description for the program to be printed
with the help output.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>