Add Cluster Inventory plugin
This is the initial release of the Cluster Inventory plugin which
collects and presents information about your cluster. It produces a
report in the Sonobuoy results format which includes details of the
following:

 * Cluster Nodes
 * Control Plane
 * CNI
 * Namespaces
 * Workloads running in each namespace:
    * pods
    * deployments
    * replicasets
    * replicationcontrollers
    * statefulsets
    * cronjobs
    * jobs

The workloads are presented as a tree. The plugin iterates over each of
the object types and forms a tree based on the controller
OwnerReferences.

The plugin can produce a report in the Sonobuoy results format or as
JSON. The Sonobuoy results format summarises some of the data obtained
to allow the results to be more easily inspected. The JSON report is
produced by marshalling each of the objects and data is presented as is.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>