Add initial commit for e2e-skeleton
Basic framework for people wanting to write their own tests
which need to interact with the k8s cluster.

Using the k/k test framework is extremely heavy since it has
a large number of dependencies which are hard to reduce. Using
the newer, lighter e2e-test-framework allows easy invocation,
easy importing, and a quick way to be able to work with the clusters.

Currently uses `raw` output but Sonobuoy will soon support gojson
output as well.

Signed-off-by: John Schnake <jschnake@vmware.com>