Plugin-helper: ResultsWriter shouldn't err if no resultsDir set
In non-cluster environments (e.g. running code that would be containerized
for a plugin, but running it locally) the SONOBUOY_RESULTS_DIR won't
be set and so we shouldn't cause users to write special if/then logic
to turn off that functionality.

We should simply detect this state and write to stdout and/or log
when appropriate.

Signed-off-by: John Schnake <jschnake@vmware.com>